requirements.txt
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
colorama==0.4.4
distlib==0.3.4
Django==4.0.4
filelock==3.6.0
iniconfig==1.1.1
packaging==21.3
platformdirs==2.5.1
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.8
pytest==7.1.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
virtualenv==20.14.1

djangorestframework~=3.13.1